backend/requirements.txt
crewai>=0.30.0
crewai-tools>=0.2.0
fastapi>=0.110.0
uvicorn>=0.29.0
python-dotenv>=1.0.0
httpx>=0.27.0
